requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2
psycopg2-binary
